+++ openexr-rs/src/rgba/rgba_file.rs
@@ -167,8 +167,8 @@ impl RgbaOutputFile {
             sys::Imf_RgbaOutputFile_setFrameBuffer(
                 self.0,
                 data.as_ptr() as *const sys::Imf_Rgba_t,
-                x_stride as u64,
-                y_stride as u64,
+                x_stride,
+                y_stride,
             )
             .into_result()?;
         }
@@ -392,8 +392,8 @@ impl RgbaInputFile {
             sys::Imf_RgbaInputFile_setFrameBuffer(
                 self.0,
                 pixels.as_mut_ptr() as *mut sys::Imf_Rgba_t,
-                x_stride as u64,
-                y_stride as u64,
+                x_stride,
+                y_stride,
             )
             .into_result()?;
         }
