+++ openexr-sys/openexr-c/include/std_string.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "openexr-api-export.h"
 
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -32,7 +33,7 @@ OPENEXR_CPPMM_API unsigned int std__string_assign(
     std_string_t * this_
     , std_string_t * * return_
     , char const * s
-    , unsigned long count);
+    , size_t count);
 #define std_string_assign std__string_assign
